refactor: move selector to prompt pkg
