+++ cmd/manifest/selector.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 
 	"github.com/okteto/okteto/cmd/utils"
-	oktetoLog "github.com/okteto/okteto/pkg/log"
+	"github.com/okteto/okteto/pkg/prompt"
 )
 
 const (
@@ -88,7 +88,7 @@ func selectDockerfiles(ctx context.Context, cwd string) ([]string, error) {
 		}
 
 		selectorItems := getDockerfilesSelectorItems(dockerfiles)
-		selector := oktetoLog.NewOktetoSelector("Do you need to build any of the following Dockerfiles as part of your development environment?", selectorItems, "")
+		selector := prompt.NewOktetoSelector("Do you need to build any of the following Dockerfiles as part of your development environment?", selectorItems, "")
 		selection, err := selector.Ask()
 		if err != nil {
 			return nil, err
@@ -126,10 +126,10 @@ func validateDockerfileSelection(dockerfiles []string) error {
 	return nil
 }
 
-func getDockerfilesSelectorItems(items []string) []oktetoLog.SelectorItem {
-	dockerfilesItems := make([]oktetoLog.SelectorItem, 0)
+func getDockerfilesSelectorItems(items []string) []prompt.SelectorItem {
+	dockerfilesItems := make([]prompt.SelectorItem, 0)
 	for _, d := range items {
-		dockerfilesItems = append(dockerfilesItems, oktetoLog.SelectorItem{
+		dockerfilesItems = append(dockerfilesItems, prompt.SelectorItem{
 			Name:   d,
 			Label:  d,
 			Enable: true,
