+++ cmd/up/up.go
@@ -33,6 +33,7 @@ import (
 	"github.com/okteto/okteto/cmd/utils/executor"
 	"github.com/okteto/okteto/pkg/analytics"
 	"github.com/okteto/okteto/pkg/discovery"
+	"github.com/okteto/okteto/pkg/prompt"
 
 	"github.com/okteto/okteto/pkg/cmd/pipeline"
 	"github.com/okteto/okteto/pkg/config"
@@ -267,7 +268,7 @@ func Up() *cobra.Command {
 				}
 
 				selectorItems := utils.GetItemsForDevSelector(oktetoManifest.Dev)
-				selector := oktetoLog.NewOktetoSelector("Select the development container you want to activate:", selectorItems, "Development container")
+				selector := prompt.NewOktetoSelector("Select the development container you want to activate:", selectorItems, "Development container")
 				dev, err = utils.SelectDevFromManifest(oktetoManifest, selector)
 				if err != nil {
 					return err
