+++ pkg/prompt/errors.go
@@ -1,0 +1,6 @@
+package prompt
+
+import "errors"
+
+// ErrInvalidOption is raised when the selector has an invalid option
+var ErrInvalidOption = errors.New("invalid option")
